Refactor Amount

+++ src/PaymentGateway.Api/Application/Mappers/BankRequestMapper.cs
@@ -17,7 +17,7 @@ public class BankRequestMapper
             CardNumber = payment.Card.Number,
             ExpiryDate = $"{payment.Card.ExpiryMonth:D2}/{payment.Card.ExpiryYear}",
             Currency = payment.Money.Currency,
-            Amount = payment.Money.Amount,
+            Amount = (int)payment.Money.Amount,
             Cvv = payment.Card.Cvv
         };
     }
